% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slides.R
\name{render_ioslides}
\alias{render_ioslides}
\title{Render ioslides from an .Rmd file}
\usage{
render_ioslides(path, ...)
}
\arguments{
\item{path}{Path to an .Rmd file.}

\item{...}{Argumets passed to [rmarkdown::render].}
}
\value{
It's called for the side effect of creating an .html-ioslides
  version of `path`.
}
\description{
Render ioslides from an .Rmd file
}
\examples{
if (interactive()) {
  in_ds_incubator <- identical(fs::path_file(here::here()), "ds-incubator")
  stopifnot(in_ds_incubator)

  render_ioslides("2020-01-21-debugging-with-rstudio.Rmd")

  message("Opening slides.html in your web browser")
  browseURL("slides.html")
}
}
